extract ManuscriptType
