+++ ebl/corpus/application/schemas.py
@@ -1,3 +1,4 @@
+from ebl.corpus.domain.manuscript_type import ManuscriptType
 from ebl.corpus.domain.provenance import Provenance
 from marshmallow import (
     Schema,
@@ -22,7 +23,6 @@ from ebl.corpus.domain.manuscript_line import ManuscriptLine
 from ebl.corpus.domain.line_variant import LineVariant
 from ebl.corpus.domain.manuscript import (
     Manuscript,
-    ManuscriptType,
     OldSiglum,
     is_invalid_non_standard_text,
     is_invalid_standard_text,
